Replace PropagatorDict using PropagatorCollections
PropagatorDict was unwieldy. It doesn't make that much sense to have a
struct containing only several Dicts that all have the same keys (rf
powers in this case) but different contents. Replace this with a single
Dict containg a struct (PropagatorCollectionRF) whose fields mirror the
contents of the Dicts that were in PropagatorDict. Several of the Dicts
contained Dicts that also share keys (pulse timings), so also separately
extract those into PropagatorCollectionTiming and put a Dict of them in
PropagatorCollectionRF.