Rework PropagationChunk creation
Instead of going through the logic for iterating through the pulse
sequence and deciding what the PropagationChunks need twice (once to
populate the pulse_cache and once to build the PropagationGenerator)
only go through it once. This is accomplished by having the
PropagatorChunks record the pulse sequence elements that they need and
populating the pulse_cache in a single pass. Actually filling in the
propagators is then just a matter of fetching them from the prop_cache
based on the recorded elements.