Fix step counting bugs in looped PropagationChunks
The first of the initial propagators should not always start at step 1.
The number of repeats for the incrementor block should be the
old_loop_cycle not the incrementor_cycle. The incrementor starting steps
should take into account the steps from the nonlooped elements, include
one less instance of the looped element, and be different for each
incrementor.