Let SimulationParameters handle temporary arrays
There are multiple places in which a series of matrix multiplications
are done. The intermediate results are not saved, so allocations can be
saved by reusing a temporary array to multiply into. Since the number of
multiplications is not fixed in some cases, which of the original arrays
holds the final result and which holds unneeded intermediate values in
not predetermined. Since there is not way for a function to alter the
binding of the variable holding the temporary in the scope that called
it, many functions had to both accept the temporary as an argument and
then also return it so the right array could be bound in the calling
scope.

By storing the temporary in a vector in the parameters it can be rebound
in one function and have that take hold in every scope. This removes the
need to explicitly pass the temporary into many functions or deal with
it being returned. The calling scope no longer needs to care about how
these functions handle the temporary. This simplifies things, especially
since the temporary was passed up and down several call chains. Now only
the SimulationParameters needed to passed into these functions.