Implement faster Block hashing/comparing
Storing Tuple{Block,Int} in a Dict by using the string representation is
an easy way to use the Block contents instead of the object ID, but it
quite slow. Collapsing the Block into an equivalent Vector{Pulse} and
using that in the hash and isequal functions is much faster.