Use 2 dimensional arrays in single modes
Using 3D arrays with CPUSingleMode and GPUSingleMode meant that mul! had
to check the size of the third dimension and if it was 1 call dropdims
on the arrays and call gemm!. The calls to dropdims added a ton of
overhead, which is avoided by using dispatch to decide to call gemm! vs
square_strided_batch_gemm!.