Avoid caching redundant Blocks
When a Block has 1 repeat and its only element is another Block it is
essentially the same as that Block. Actually constructing such a Block
will result in a Block of higher rank and then the same thing gets
cached in multiple places. This redundancy can be avoided by not
constructing a new Block in such cases and just using the original
Block.