Rework generation of pulse propagators
Instead of saving a set of combined propagators that can be used to
iterate pulse propagators to the next gamma angle, generate all of the
pulses up front. This approach uses a set of combined propagators that
all start at the first step to generate the pulses.

This should reduce some duplication of work in cases where the same
pulse is used on different gamma angles. In terms of memory use, this
change requires keeping a propagator for each pulse for each gamma
angle, but no longer requires keeping the combined propagators. Overall
this means less memory used in cases where a set of combined propagators
with two propagators per gamma angle was only used for a single pulse,
but more memory in cases where a set of combined propagators could be
used for several pulses.